% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_qry4ms}
\alias{run_qry4ms}
\title{Run the qry4ms Application}
\usage{
run_qry4ms()
}
\description{
Run the qry4ms Application locally. Shiny deployment
\href{https://plyush1993.shinyapps.io/qry4ms/}{here}.
}
